+++ photo_list.php
@@ -13,49 +13,52 @@ require_once("inc/Utilities/PhotoDAO.class.php");
 
 session_start();
 
-// View a list of photo, and click link to upload photo
 
-//Initializing the DAO's
-UserDAO::init();
-PhotoDAO::init();
+if(isset($_SESSION) && isset($_SESSION["loggedin"])) {
+    // View a list of photo, and click link to upload photo
 
-$u = UserDAO::getUser("".$_SESSION['loggedin']);
+    //Initializing the DAO's
+    UserDAO::init();
+    PhotoDAO::init();
 
-$Uid = $u->getId();
+    $u = UserDAO::getUser("".$_SESSION['loggedin']);
+
+    $Uid = $u->getId();
 
 
-Page::$subTitle = $u->getFirstName()." ".$u->getLastName()." : My Photos"; //TODO: make it a better name?
-Page::header("PhotoGallery");
-$photo_list =PhotoDAO::getPhoto($Uid);
-if (isset($_GET["action"]) && $_GET["action"] == "delete")  {
-    //Use the DAO to delete the corresponding registration
-    PhotoDAO::deletePhoto($_GET['PhotoID']);
+    Page::$subTitle = $u->getFirstName()." ".$u->getLastName()." : My Photos";
+    Page::header("PhotoGallery");
     $photo_list =PhotoDAO::getPhoto($Uid);
-    }
-if (isset($_GET["action"]) && $_GET["action"] == "edit")  {
-    $newphotodetails = new Photo;
-
-    Page::editPhotoForm($photo_list);
-    if (isset($_POST["action"]) && $_POST["action"] == "edit")  {
-        $newphotodetails->setDisplay_name($_POST['display_name']);
-        $newphotodetails->setDescription($_POST['description']);
-        $newphotodetails->setId($_POST['Photoid']);
+    if (isset($_GET["action"]) && $_GET["action"] == "delete")  {
+        //Use the DAO to delete the corresponding registration
+        PhotoDAO::deletePhoto($_GET['PhotoID']);
         $photo_list =PhotoDAO::getPhoto($Uid);
-//get the photos based on the user who signed in
-PhotoDAO::updatePhoto($newphotodetails);
+        }
+    if (isset($_GET["action"]) && $_GET["action"] == "edit")  {
+        $newphotodetails = new Photo;
 
-
-    }   
-    if (isset($_POST['submit'])) { 
-        Page::$subTitle="Updated Successfully!";
-        Page::header("PhotoGallery");
-        Page::backToPhotoList();
+        Page::editPhotoForm($photo_list);
+        if (isset($_POST["action"]) && $_POST["action"] == "edit")  {
+            $newphotodetails->setDisplay_name($_POST['display_name']);
+            $newphotodetails->setDescription($_POST['description']);
+            $newphotodetails->setId($_POST['Photoid']);
+            $photo_list =PhotoDAO::getPhoto($Uid); //get the photos based on the user who signed in
+            PhotoDAO::updatePhoto($newphotodetails);
+        }   
+        if (isset($_POST['submit'])) { 
+            Page::$subTitle="Updated Successfully!";
+            Page::header("PhotoGallery");
+            Page::backToPhotoList();
+        }
+        Page::footer();
     }
+    else{
+        //get the photos based on the user who signed in
+        Page::displayPhotolist($photo_list);
+        Page::footer();
+    }
+}else{
+    Page::$subTitle = "Please login first";
+    Page::header("Access Denied");
     Page::footer();
-}
-else{
-
-//get the photos based on the user who signed in
-Page::displayPhotolist($photo_list);
-Page::footer();
 }
